Made add dog component accessible through adopted, training and health pages
